Refactor blog architecture with modular, performant, and reactive design
