+++ blogs.html
@@ -10,12 +10,7 @@
 
 
     <main id="main-content">
-        <div class="centered-content">
-            <section id="blog">
-                <h1 data-key="blog-title">Blog</h1>
-                <ul id="blog-list"></ul>
-            </section>
-        </div>
+        <!-- Content will be dynamically loaded by the new architecture -->
     </main>
 
     <script src="https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js"></script>
