+++ css/styles.css
@@ -171,3 +171,163 @@ section {
     margin: 0 auto;
     padding: 0 20px;
 }
+
+/* Loading Spinner Styles */
+.loading-spinner {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding: 40px;
+    text-align: center;
+}
+
+.spinner {
+    border: 4px solid #f3f3f3;
+    border-top: 4px solid #3498db;
+    border-radius: 50%;
+    width: 40px;
+    height: 40px;
+    animation: spin 1s linear infinite;
+    margin-bottom: 20px;
+}
+
+@keyframes spin {
+    0% { transform: rotate(0deg); }
+    100% { transform: rotate(360deg); }
+}
+
+body.dark-mode .spinner {
+    border: 4px solid #333;
+    border-top: 4px solid #61dafb;
+}
+
+/* Error Display Styles */
+.error-display {
+    padding: 20px;
+    text-align: center;
+    background: #ffe6e6;
+    border: 1px solid #ffcccc;
+    border-radius: 5px;
+    margin: 20px;
+}
+
+body.dark-mode .error-display {
+    background: #3d1a1a;
+    border-color: #5c2e2e;
+    color: #ffaaaa;
+}
+
+.error-content h2 {
+    color: #d32f2f;
+    margin-bottom: 10px;
+}
+
+body.dark-mode .error-content h2 {
+    color: #ff6b6b;
+}
+
+.error-message {
+    margin: 15px 0;
+    font-family: 'Fira Code', monospace;
+    font-size: 0.9em;
+}
+
+.retry-button {
+    background: #3498db;
+    color: white;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 14px;
+    margin-top: 10px;
+}
+
+.retry-button:hover {
+    background: #2980b9;
+}
+
+body.dark-mode .retry-button {
+    background: #61dafb;
+    color: #000;
+}
+
+body.dark-mode .retry-button:hover {
+    background: #4fa8c5;
+}
+
+/* Content List Styles */
+.content-section {
+    margin-bottom: 30px;
+}
+
+.content-list-container {
+    margin-top: 20px;
+}
+
+.content-list {
+    list-style: none;
+    padding: 0;
+}
+
+.content-item {
+    margin-bottom: 15px;
+    padding: 15px;
+    border: 1px solid #e0e0e0;
+    border-radius: 5px;
+    transition: background-color 0.3s ease;
+}
+
+.content-item:hover {
+    background-color: #f9f9f9;
+}
+
+body.dark-mode .content-item {
+    border-color: #444;
+    background-color: #333;
+}
+
+body.dark-mode .content-item:hover {
+    background-color: #404040;
+}
+
+.content-link {
+    color: #3498db;
+    text-decoration: none;
+    font-weight: 500;
+    font-size: 1.1em;
+}
+
+.content-link:hover {
+    text-decoration: underline;
+}
+
+body.dark-mode .content-link {
+    color: #61dafb;
+}
+
+/* Loading State for Content */
+.loading {
+    text-align: center;
+    padding: 40px;
+    font-style: italic;
+    opacity: 0.7;
+}
+
+/* Improved responsive design */
+@media (max-width: 600px) {
+    .content-item {
+        margin-bottom: 10px;
+        padding: 10px;
+    }
+    
+    .loading-spinner {
+        padding: 20px;
+    }
+    
+    .spinner {
+        width: 30px;
+        height: 30px;
+    }
+}
