+++ index.html
@@ -19,15 +19,7 @@
 
     <!-- Main Content -->
     <main id="main-content">
-        <div class="centered-content">
-            <!-- Your page content here -->
-            <h1>Sebastian Cubides</h1>
-            <h2>A website's whitepaper.</h2>
-
-            <p>
-                This is a website's whitepaper.
-            </p>
-        </div>
+        <!-- Content will be dynamically loaded by the new architecture -->
     </main>
 
     <!-- Scripts -->
