+++ projects.html
@@ -8,12 +8,7 @@
 </head>
 <body>
     <main id="main-content">
-        <div class="centered-content">
-            <section id="projects">
-                <h1 data-key="projects-title">Projects</h1>
-                <ul id="project-list"></ul>
-            </section>
-        </div>
+        <!-- Content will be dynamically loaded by the new architecture -->
     </main>
     <script src="https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"></script>
